Move functions to new places
